Metadata columns
Added new table according to the proposal with the new metadata columns
in the package de.bund.bfr.knime.pmm.extendedtable. There is two new
metadata cells: TimeSeriesMetadata and ModelMetadata. The former only
keeps one kind of literature unlike the latter one that keeps literature
of the model and the estimated model. Both of them keeps the matrix and
the organism.

The PMF Reader uses these new tables though the old ones are still in
the old package since the rest of the nodes rely on them.