FSK: Added library registry (FSKNodePlugin.LibRegistry)
Handles R libraries so that they do not need to be installed every time
an FSK node is run. It keeps an installation path with uncompressed R
libraries and secondary path with R binary libraries ready to be bundled
into an FSK file.

FSKX Creator is the only node (so far) that uses it. The first time the
creator runs it retrieves all the libraries and their dependencies as
usual. The next time though, the libraries are retrieved directly from
disk instead of downloading/installing them again.