Bugfix with non valid unit ids
When multiple illegal unit ids need to be converted to valid SIds in
SBML if they are similar they produce the same id after conversion which
can not be added to the model metadata more than once.

For example the units ºC and °C would be after conversion C resulting
into error the second time the node tries to save the unit.